requirements.txt
python-dotenv==1.0.1
azure-ai-documentintelligence==1.0.0b3
langchain==0.1.20
tqdm==4.66.4
pandas==2.2.2
openai>=1.12.0
azure-identity==1.16.1
mysql-connector-python==8.2.0 